requirements.txt
django==5.1.4
gunicorn==23.0.0
whitenoise==6.9.0
# dotenv==0.9.9
pytest-cov==6.1.1
pytest-mock==3.14.1